security: implement all fixes from AUDIT_2026_05_26_B
HIGH
- OAuth CSRF: add HMAC-signed state param to Strava+Garmin auth URLs,
  verify on callback; new lib/oauth-state.ts utility
- encryptIfNeeded: replace fragile colon-count heuristic with strict hex regex
- coach/chat approvedAction.toolName: z.enum([...]) instead of z.string()

MEDIUM
- auth.ts: login rate limit 10 attempts / 600 s per email via checkRateLimit
- coach/chat: chat rate limit 10 req / 60 s per user
- sports POST kind=type: verify sportId ownership before creating WorkoutType
- planner templates POST: verify sportId + typeId ownership
- planner workouts GET: validate from/to date params with /^\d{4}-\d{2}-\d{2}$/
- ChatInterface: wrap sendWithPayload in useCallback to fix stale closure;
  fix send deps to use sendWithPayload directly

LOW
- profile route: yearsTraining → z.coerce.number().int()
- backfill routes: parseInt radix 10 + || 30 fallback for NaN
- workouts PATCH color: add /^#[0-9a-fA-F]{6}$/ regex validation
- callbacks: use NEXTAUTH_URL env var as redirect URI base (host-header safe)
- fitness/cache: first-run zone default uses computed maxHR/restHR not 185/45

+++ app/api/settings/profile/route.ts
@@ -13,7 +13,7 @@ const schema = z.object({
   maxHeartRate:    z.coerce.number().min(100).max(230).optional().nullable(),
   restingHeartRate:z.coerce.number().min(20).max(100).optional().nullable(),
   primaryGoal:     z.string().max(200).optional().nullable(),
-  yearsTraining:   z.coerce.number().min(0).max(80).optional().nullable(),
+  yearsTraining:   z.coerce.number().int().min(0).max(80).optional().nullable(),
 });
 
 export async function POST(req: NextRequest) {
